Split out SpectralWindow class to its own module
This shrinks the big blob that is dataset.py a bit, and makes the
use of SpectralWindow by internal modules like applycal a bit cleaner.
