+++ katdal/test/test_spectral_window.py
